Mark attachments added after unpublishing as draft in Asset Manager
Currently, after a document is unpublished, any newly added attachments are marked as 'not draft' in Asset Manager.  This means it could be possible for the attachments to be accessed by anyone by guessing the URL.

This changes the logic to ensure that attachments added after unpublishing are marked as draft in Asset Manager, while retaining the correct status of attachments added before unpublishing.

https://trello.com/c/PlPb1iNY/337-fix-attachments-on-unpublished