The announcement finder now uses the new rummager search endpoint
Finders in Whitehall used the '/advanced_search' endpoint in
Rummager. This does not surface all relevant content because
it only uses the goverment index.

This change makes the annoncement finder use the newer /search
endpoint that spans the government and govuk search indices and
should therefore be capable of returning all relevant results.